requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
alluka==0.4.0
attrs==24.3.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ciso8601==2.3.2
colorlog==6.9.0
frozenlist==1.5.0
greenlet==3.1.1
hikari==2.1.0
hikari-tanjun==2.17.7
idna==3.10
multidict==6.1.0
nest-asyncio==1.6.0
orjson==3.10.15
playwright==1.50.0
propcache==0.2.1
pycares==4.5.0
pycparser==2.22
pyee==12.1.1
python-dotenv==1.0.1
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
